Fancy racer filters
Use bootstrap-slider for selecting season range and checkboxes for race
classes, and jQuery to update the table.  To support this, return results
from a json api point.

This also adds an explicit checkbox, only in admin-mode, to show racers that
have no races.  The idea is that admins care about such racers because
they've probably recently added them and are entering races, but the world
doesn't need to see those (yet).